refactor(react-ds-global-form): bindField → own folder, .ts + createElement, split types
Drive-by tidy of the field-composition factory:

- `common/bindField.tsx` → `common/bindField/index.ts`. Authored with
  `React.createElement` instead of JSX, so it is a plain `.ts` module (no JSX
  runtime in a non-component factory file). Behaviour is identical — every
  field is bound through it and the full suite is unchanged.
- Types (`BindMode`, `BindFieldOptions`, `FieldBindingProps`) moved to
  `common/bindField/types.ts` and re-exported from the index.
- Repoint all importers to `#lib/common/bindField/index.js` and the
  `common/index.ts` barrel + Wrapper story to the folder path.

git mv preserves history. Gate: tsc clean; vitest 59 files / 212 pass / 1 skip /
4 todo; `bun run check` passes; `build:storybook` succeeds.
