+++ packages/react/ds-global-form/src/lib/common/Wrapper/Wrapper.stories.tsx
@@ -6,7 +6,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 import * as decorators from "storybook/decorators.js";
 import { TextareaInput } from "#lib/subcomponent/TextareaInput/index.js";
 import { TextInput } from "#lib/subcomponent/TextInput/index.js";
-import bindField from "../bindField.js";
+import bindField from "../bindField/index.js";
 import Component from "./Wrapper.js";
 
 // The Wrapper renders a *bound* component (one that consumes name/registerProps).
