+++ packages/react/ds-global-form/src/lib/component/ColorField/ColorField.tsx
@@ -1,4 +1,4 @@
-import bindField from "#lib/common/bindField.js";
+import bindField from "#lib/common/bindField/index.js";
 import withWrapper from "#lib/common/Wrapper/withWrapper.js";
 import { ColorInput } from "#lib/subcomponent/ColorInput/index.js";
 import type { ColorFieldProps } from "./types.js";
