+++ packages/react/ds-global-form/src/lib/component/DateTimeInputField/DateTimeInputField.tsx
@@ -1,4 +1,4 @@
-import bindField from "#lib/common/bindField.js";
+import bindField from "#lib/common/bindField/index.js";
 import withWrapper from "#lib/common/Wrapper/withWrapper.js";
 import { DateTimeInput } from "#lib/subcomponent/DateTimeInput/index.js";
 import type { DateTimeInputFieldProps } from "./types.js";
