+++ packages/react/ds-global-form/src/lib/component/FileUploadField/FileUploadField.tsx
@@ -1,4 +1,4 @@
-import bindField from "#lib/common/bindField.js";
+import bindField from "#lib/common/bindField/index.js";
 import withWrapper from "#lib/common/Wrapper/withWrapper.js";
 import { FileUploadInput } from "#lib/subcomponent/FileUploadInput/index.js";
 import type { FileUploadFieldProps } from "./types.js";
