+++ packages/react/ds-global-form/src/lib/component/HiddenField/HiddenField.tsx
@@ -1,4 +1,4 @@
-import bindField from "#lib/common/bindField.js";
+import bindField from "#lib/common/bindField/index.js";
 import { InvisibleWrapper, withWrapper } from "#lib/common/Wrapper/index.js";
 import { HiddenInput } from "#lib/subcomponent/HiddenInput/index.js";
 import type { HiddenFieldProps } from "./types.js";
