+++ packages/react/ds-global-form/src/lib/component/NumberField/NumberField.tsx
@@ -1,6 +1,6 @@
 // Generated by @canonical/summon-component:react v0.1.0
 
-import bindField from "#lib/common/bindField.js";
+import bindField from "#lib/common/bindField/index.js";
 import withWrapper from "#lib/common/Wrapper/withWrapper.js";
 import { NumberInput } from "#lib/subcomponent/NumberInput/index.js";
 import type { NumberFieldProps } from "./types.js";
