+++ packages/react/ds-global-form/src/lib/component/PasswordField/PasswordField.tsx
@@ -1,6 +1,6 @@
 // Generated by @canonical/summon-component:react v0.1.0
 
-import bindField from "#lib/common/bindField.js";
+import bindField from "#lib/common/bindField/index.js";
 import withWrapper from "#lib/common/Wrapper/withWrapper.js";
 import { PasswordInput } from "#lib/subcomponent/PasswordInput/index.js";
 import type { PasswordFieldProps } from "./types.js";
